Rename assets money
